refactor: Edit bottomsheet directory
